requirements.txt
argon2-cffi==23.1.0
cryptography==42.0.5
pycryptodome==3.20.0
